Move sim_response computations into one function
Calcuating the simulated response matrix in add_{pi, probs,
quantile}.glm was done in each of these functions using the same
code. This made it very annoying to fix changes regarding this
function. I placed one function in add_pi.glm that does the
computation for us. The function is then called by the three scripts
individually when needed.