Changed pedicted value for binomial regression is E(X) from pHat
Previous functionality was to populate the column, "pred" with the estimated value from `predict`. Even for binomial models, `predict` reports the estimated value for p rather than the expected value of the response. This change makes it so add_pi, add_probs, and add_quantile will generate estimates of the response. A warning is also added to alert users of the change from the typical default.

Note that this does not change the naming convention for this column (default name "pred").