Changes related to binomial regressio in add_probs.glm
This commit fixes the same typo in the previous commit in add_probs as well as a few other changes related to error messages. First, I rearranged some of the errors and warnings in add_probs.glm to fail faster. I also had to rearrange a conditional statement to ensure that hte right function (sim_probs_logistic vice sim_probs_other) was called depending on whether we did a binomial vice Bernoulli RV.