add_ci Method for Negative Binomial Fits.
add_ci.glm can handle quasipoisson regression, which is one method of
dealing with over-dispersion in ordinal data, but another method is
using Negative Binomial distribution. The distributional assumptions
are similar, but the variance of the conditional response
distributions differ: In quasipoisson regression, the variance is a
linear function of the mean, but in NB regression, the variance is a
quadratic function of the mean. Fitting NB regressions in R is usually
done with glm.nb(), not glm(), and produce an object of class negbin.

In this commit, I add a parametric ci method and a bootstrap ci
method. These changes are not substantial. I had planned on adding a
PI method for NB regression, but I now see that it will require some
more work to get these intervals -- arm::sim does not (and probably
will never) handle negbin objects. I hope to eventually have a NB PI
method available, because it seems that the PI method for QP
regressions does not have correct coverage probability.

McCullagh and Nelder's GLMs book (p. 199) has some info about NB
models.