Adding Rdpack Dependency
Rdpack is a package that uses RdMacros to insert references from a
Bibtex file. I copied the Bibtex file from the lit-review to the inst/
folder for testing. Hopefully this will make adding a reference to
each function easy.

My plan is to see if this actuals works as it
claims (https://cran.r-project.org/web/packages/Rdpack/vignettes/Inserting_bibtex_references.pdf),
then just try to document each method. I don't think we should place
references in the documentation of the generics.