add_quantile.glm changed to via simulation
The quant_gaussian function works if you use the identity link but fails if you have either log or inverse links (both supported by glm). Two options:  One is enable quant_gaussian only for identity link. The other is to just have it use the same simulation methods that we use for other glm families. I opted for the latter. Maybe look to change in the future?